source/requirements.txt
oauth2client
google-api-python-client
google-api-core
google-auth
google-auth-oauthlib
google-auth-httplib2
